added cinema halls management for admin

+++ src/main/java/com/example/cinema_reservation/serviceimpl/CinemaHallServiceImpl.java
@@ -1,8 +1,11 @@
 package com.example.cinema_reservation.serviceimpl;
 
 import com.example.cinema_reservation.model.CinemaHall;
+import com.example.cinema_reservation.model.Screening;
 import com.example.cinema_reservation.repository.CinemaHallRepository;
+import com.example.cinema_reservation.repository.ScreeningRepository;
 import com.example.cinema_reservation.service.CinemaHallService;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,10 +16,13 @@ import java.util.Optional;
 public class CinemaHallServiceImpl implements CinemaHallService {
 
     private final CinemaHallRepository cinemaHallRepository;
+    private final ScreeningRepository screeningRepository;
 
     @Autowired
-    public CinemaHallServiceImpl(CinemaHallRepository cinemaHallRepository) {
+    public CinemaHallServiceImpl(CinemaHallRepository cinemaHallRepository,
+                                 ScreeningRepository screeningRepository) {
         this.cinemaHallRepository = cinemaHallRepository;
+        this.screeningRepository = screeningRepository;
     }
 
     public List<CinemaHall> findAllCinemaHalls() {
@@ -25,6 +31,10 @@ public class CinemaHallServiceImpl implements CinemaHallService {
 
     public Optional<CinemaHall> findCinemaHallById(Long id) {
         return cinemaHallRepository.findById(id);
+    }
+
+    public Optional<CinemaHall> findCinemaHallByName(String name) {
+        return cinemaHallRepository.findByName(name);
     }
 
     public CinemaHall saveCinemaHall(CinemaHall cinemaHall) {
@@ -43,12 +53,25 @@ public class CinemaHallServiceImpl implements CinemaHallService {
                 .orElseThrow(() -> new RuntimeException("Cinema hall not found with id: " + id));
     }
 
+    @Override
+    @Transactional
     public void deleteCinemaHallById(Long id) {
-        cinemaHallRepository.deleteById(id);
+        Optional<CinemaHall> optionalCinemaHall = cinemaHallRepository.findById(id);
+
+        if (optionalCinemaHall.isPresent()) {
+            List<Screening> screenings = screeningRepository.findAllByCinemaHallId(id);
+
+            if (!screenings.isEmpty()) {
+                screeningRepository.deleteAll(screenings);
+            }
+            cinemaHallRepository.deleteById(id);
+        }
     }
 
-    @Override
-    public int cinemaHallCount() {
-        return (int) cinemaHallRepository.count();
+        @Override
+        public int cinemaHallCount () {
+            return (int) cinemaHallRepository.count();
+        }
     }
-}
+
+
